[Tests] Use consistent strategy for test manifest lists.
 - This moves to using one test list per test suite, and then keeping all of the
   test list within that suite organized adjacent to the test suite (in a hope
   this will make maintenance easier).

 - Also, applied a couple renames to make the test file names more consistent
   with the test case inside.

 - Also, added a number of missing tests to the lists.
