+++ Tests/Get/XCTestManifests.swift
@@ -13,9 +13,9 @@ import XCTest
 #if !os(OSX)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(ModuleDependencyTests.allTests),
-        testCase(PrimitiveResolutionTests.allTests),
-        testCase(ValidSourcesTests.allTests),
+        testCase(GetTests.allTests),
+        testCase(GitTests.allTests),
+        testCase(VersionGraphTests.allTests),
     ]
 }
 #endif
