+++ Tests/LinuxMain.swift
@@ -5,44 +5,26 @@
 
 import XCTest
 
-@testable import TransmuteTestSuite
-@testable import UtilityTestSuite
-@testable import FunctionalTestSuite
-@testable import GetTestSuite
-@testable import ManifestSerializerTestSuite
-@testable import PackageDescriptionTestSuite
-@testable import PackageTypeTestSuite
-@testable import BuildTestSuite
+import BuildTestSuite
+import FunctionalTestSuite
+import GetTestSuite
+import ManifestSerializerTestSuite
+import OptionsParserTestSuite
+import PackageDescriptionTestSuite
+import PackageTypeTestSuite
+import PkgConfigTestSuite
+import TransmuteTestSuite
+import UtilityTestSuite
 
-XCTMain([
-    testCase(TestClangModulesTestCase.allTests),
-    testCase(DependencyResolutionTestCase.allTests),
-    testCase(FileTests.allTests),
-    testCase(GetTests.allTests),
-    testCase(GitTests.allTests),
-    testCase(InvalidLayoutsTestCase.allTests),
-    testCase(ManifestTests.allTests),
-    testCase(MiscellaneousTestCase.allTests),
-    testCase(ManifestSerializerTestSuite.PackageTests.allTests),
-    testCase(ManifestSerializerTestSuite.JSONSerializationTests.allTests),
-    testCase(ModuleDependencyTests.allTests),
-    testCase(ValidSourcesTests.allTests),
-    testCase(PrimitiveResolutionTests.allTests),
-    testCase(PackageDescriptionTestSuite.PackageTests.allTests),
-    testCase(PackageTypeTestSuite.PackageTests.allTests),
-    testCase(PathTests.allTests),
-    testCase(RelativePathTests.allTests),
-    testCase(ShellTests.allTests),
-    testCase(StatTests.allTests),
-    testCase(StringTests.allTests),
-    testCase(TOMLTests.allTests),
-    testCase(URLTests.allTests),
-    testCase(ValidLayoutsTestCase.allTests),
-    testCase(VersionGraphTests.allTests),
-    testCase(VersionTests.allTests),
-    testCase(WalkTests.allTests),
-    testCase(ModuleMapsTestCase.allTests),
-    testCase(DescribeTests.allTests),
-    testCase(GitUtilityTests.allTests),
-    testCase(PackageVersionDataTests.allTests),
-])
+var tests = [XCTestCaseEntry]()
+tests += BuildTestSuite.allTests()
+tests += FunctionalTestSuite.allTests()
+tests += GetTestSuite.allTests()
+tests += ManifestSerializerTestSuite.allTests()
+tests += OptionsParserTestSuite.allTests()
+tests += PackageDescriptionTestSuite.allTests()
+tests += PackageTypeTestSuite.allTests()
+tests += PkgConfigTestSuite.allTests()
+tests += TransmuteTestSuite.allTests()
+tests += UtilityTestSuite.allTests()
+XCTMain(tests)
