+++ Tests/PkgConfig/XCTestManifests.swift
@@ -13,9 +13,7 @@ import XCTest
 #if !os(OSX)
 public func allTests() -> [XCTestCaseEntry] {
     return [
-        testCase(ModuleDependencyTests.allTests),
-        testCase(PrimitiveResolutionTests.allTests),
-        testCase(ValidSourcesTests.allTests),
+        testCase(PkgConfigParserTests.allTests),
     ]
 }
 #endif
