+++ Tests/Transmute/ModuleDependencyTests.swift
@@ -156,7 +156,7 @@ class ModuleDependencyTests: XCTestCase {
     }
 }
 
-extension Module {
+private extension Module {
     private func depends(on target: Module) {
         dependencies.append(target)
     }
@@ -164,5 +164,18 @@ extension Module {
     private var recursiveDeps: [Module] {
         sort(self)
         return dependencies
+    }
+}
+
+extension ModuleDependencyTests {
+    static var allTests : [(String, (ModuleDependencyTests) -> () throws -> Void)] {
+        return [
+           ("test1", test1),
+           ("test2", test2),
+           ("test3", test3),
+           ("test4", test4),
+           ("test5", test5),
+           ("test6", test6),
+        ]
     }
 }
