+++ Tests/Utility/PathTests.swift
@@ -78,6 +78,20 @@ class PathTests: XCTestCase {
     }
 }
 
+extension PathTests {
+    static var allTests : [(String, (PathTests) -> () throws -> Void)] {
+        return [
+                   ("test", test),
+                   ("testPrecombined", testPrecombined),
+                   ("testExtraSeparators", testExtraSeparators),
+                   ("testEmpties", testEmpties),
+                   ("testNormalizePath", testNormalizePath),
+                   ("testJoinWithAbsoluteReturnsLastAbsoluteComponent", testJoinWithAbsoluteReturnsLastAbsoluteComponent),
+                   ("testParentDirectory", testParentDirectory),
+        ]
+    }
+}
+
 class WalkTests: XCTestCase {
 
     func testNonRecursive() {
@@ -154,6 +168,17 @@ class WalkTests: XCTestCase {
     }
 }
 
+extension WalkTests {
+    static var allTests : [(String, (WalkTests) -> () throws -> Void)] {
+        return [
+                   ("testNonRecursive", testNonRecursive),
+                   ("testRecursive", testRecursive),
+                   ("testSymlinksNotWalked", testSymlinksNotWalked),
+                   ("testWalkingADirectorySymlinkResolvesOnce", testWalkingADirectorySymlinkResolvesOnce),
+        ]
+    }
+}
+
 class StatTests: XCTestCase {
 
     func test_isdir() {
@@ -195,6 +220,16 @@ class StatTests: XCTestCase {
     }
 }
 
+extension StatTests {
+    static var allTests : [(String, (StatTests) -> () throws -> Void)] {
+        return [
+                   ("test_isdir", test_isdir),
+                   ("test_isfile", test_isfile),
+                   ("test_realpath", test_realpath),
+                   ("test_basename", test_basename),
+        ]
+    }
+}
 
 class RelativePathTests: XCTestCase {
 
@@ -221,5 +256,17 @@ class RelativePathTests: XCTestCase {
         XCTAssertEqual("/1/2/3", Path.join("/1/2/4", "../3").normpath)
         XCTAssertEqual("/1/2", Path.join("/1/2/3", "..").normpath)
         XCTAssertEqual("2", Path.join("2/3", "..").normpath)
+    }
+}
+
+extension RelativePathTests {
+    static var allTests : [(String, (RelativePathTests) -> () throws -> Void)] {
+        return [
+                   ("testAbsolute", testAbsolute),
+                   ("testRelative", testRelative),
+                   ("testMixed", testMixed),
+                   ("testRelativeCommonSubprefix", testRelativeCommonSubprefix),
+                   ("testCombiningRelativePaths", testCombiningRelativePaths)
+        ]
     }
 }
